# socratec_openapi.yaml
openapi: 3.1.0
info:
  title: Socratec Traccar Extensions
  description: Socratec-specific extensions to the Traccar GPS tracking server API. This includes additional endpoints and models for enhanced trip management functionality.
  contact:
    name: Socratec Telematic GmbH
    url: https://www.socratec.de/
    email: support@socratec.de
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: '1.0.0'
servers:
  - url: http://{host}:{port}/api
    description: Socratec Server
    variables:
      host:
        default: localhost
      port:
        enum:
          - '8082'
          - '80'
        default: '8082'
security:
  - BasicAuth: []
  - ApiKey: []
...
paths:
  /trips:
    get:
      summary: Fetch a list of Trips
      description: Without params, it returns a list of Trips the user has access to
      tags:
        - Trips
      parameters:
        - name: all
          in: query
          description: Can only be used by admins or managers to fetch all entities
          schema:
            type: boolean
        - name: userId
          in: query
          description: Standard users can use this only with their own _userId_
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Trip'
        '400':
          description: No permission
          content: {}
    post:
      summary: Create a Trip (Not Supported)
      description: Trip creation is not supported. This endpoint returns HTTP 405 Method Not Allowed.
      tags:
        - Trips
      responses:
        '405':
          description: Method Not Allowed - Trip creation is not supported
          content: {}
...
  /logbook:
    post:
      summary: Create single LogbookEntry (Service Users Only)
      description: |
        Creates a single logbook entry in the system. This endpoint is restricted to service users only
        and requires authentication with a service account token. Regular users will receive a 403 Forbidden response.
        
        Service users are authenticated using the Bearer token scheme with a service account token
        configured via the `web.serviceAccountToken` configuration property.
        
        The endpoint validates required fields and automatically calculates duration if not provided.
        All creation actions are logged for audit purposes.
      tags:
        - Reports
      security:
        - ApiKey: []
      requestBody:
        required: true
        description: |
          LogbookEntry object to create. The ID field will be auto-generated and should not be provided.
          Required fields: deviceId, driverId, startPositionId, endPositionId, startTime, endTime, hash. 
          Duration will be calculated automatically if not provided.
        content:
          application/json:
            schema:
              type: object
              properties:
                deviceId:
                  type: integer
                  format: int64
                  description: ID of the device for this trip (required)
                  minimum: 1
                driverId:
                  type: integer
                  format: int64
                  description: ID of the driver who performed this trip (required)
                  minimum: 1
                startPositionId:
                  type: integer
                  format: int64
                  description: ID of the position record at trip start (required)
                  minimum: 1
                endPositionId:
                  type: integer
                  format: int64
                  description: ID of the position record at trip end (required)
                  minimum: 1
                startTime:
                  type: string
                  format: date-time
                  description: Trip start time in ISO 8601 format (required)
                endTime:
                  type: string
                  format: date-time
                  description: Trip end time in ISO 8601 format (required)
                hash:
                  type: string
                  description: Hash value for the trip (required for deduplication/integrity)
                  minLength: 1
                distance:
                  type: number
                  format: double
                  description: Total distance of the trip in meters
                  minimum: 0
                averageSpeed:
                  type: number
                  format: double
                  description: Average speed during the trip in km/h
                  minimum: 0
                maxSpeed:
                  type: number
                  format: double
                  description: Maximum speed reached during the trip in km/h
                  minimum: 0
                spentFuel:
                  type: number
                  format: double
                  description: Amount of fuel consumed during the trip in liters
                  minimum: 0
                startOdometer:
                  type: number
                  format: double
                  description: Odometer reading at trip start
                  minimum: 0
                endOdometer:
                  type: number
                  format: double
                  description: Odometer reading at trip end
                  minimum: 0
                startLat:
                  type: number
                  format: double
                  description: Latitude coordinate at trip start
                  minimum: -90
                  maximum: 90
                startLon:
                  type: number
                  format: double
                  description: Longitude coordinate at trip start
                  minimum: -180
                  maximum: 180
                endLat:
                  type: number
                  format: double
                  description: Latitude coordinate at trip end
                  minimum: -90
                  maximum: 90
                endLon:
                  type: number
                  format: double
                  description: Longitude coordinate at trip end
                  minimum: -180
                  maximum: 180
                startAddress:
                  type: string
                  maxLength: 512
                  description: Human-readable address at trip start location
                  nullable: true
                endAddress:
                  type: string
                  maxLength: 512
                  description: Human-readable address at trip end location
                  nullable: true
                duration:
                  type: integer
                  format: int64
                  description: Trip duration in milliseconds (auto-calculated if not provided)
                  minimum: 0
                notes:
                  type: string
                  description: Additional notes or comments for the trip
                  nullable: true
                type:
                  $ref: '#/components/schemas/LogbookEntryType'
              required:
                - deviceId
                - driverId
                - startPositionId
                - endPositionId
                - startTime
                - endTime
                - hash
            example:
              deviceId: 123
              driverId: 1
              startPositionId: 1
              endPositionId: 2
              startTime: "2024-11-07T08:00:00.000Z"
              endTime: "2024-11-07T10:30:00.000Z"
              distance: 125.7
              averageSpeed: 65.3
              maxSpeed: 95.0
              startLat: 52.520008
              startLon: 13.404954
              endLat: 52.370216
              endLon: 13.404954
              startAddress: "Berlin, Alexanderplatz, Germany"
              endAddress: "Berlin, Tempelhof Airport, Germany"
              notes: "Business trip to client meeting"
              type: 1
              hash: "a1b2c3d4e5f6789012345abcdef67890"
      responses:
        '200':
          description: |
            LogbookEntry successfully created. Returns the created LogbookEntry object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogbookEntry'
              example:
                id: 456
                deviceId: 123
                startTime: "2024-11-07T08:00:00.000Z"
                endTime: "2024-11-07T10:30:00.000Z"
                duration: 9000000
                distance: 125.7
                averageSpeed: 65.3
                maxSpeed: 95.0
                startLat: 52.520008
                startLon: 13.404954
                endLat: 52.370216
                endLon: 13.404954
                startAddress: "Berlin, Alexanderplatz, Germany"
                endAddress: "Berlin, Tempelhof Airport, Germany"
                notes: "Business trip to client meeting"
                type: 1
        '400':
          description: |
            Bad request. This can occur when:
            - Entry is null
            - Required fields are missing
            - Validation errors (e.g., endTime before startTime)
            - Invalid field values
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Missing required hash."
        '401':
          description: Unauthorized - Authentication required
          content: {}
        '403':
          description: |
            Forbidden - Access denied. This endpoint is restricted to service users only.
            Regular users and administrators cannot create logbook entries through this endpoint.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Access denied: Service account required"
        '500':
          description: Internal server error
          content: {}
...
components:
  schemas:
    Trip:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the trip
        deviceId:
          type: integer
          format: int64
          description: ID of the device that performed this trip
        distance:
          type: number
          format: double
          description: Total distance of the trip in meters
        averageSpeed:
          type: number
          format: double
          description: Average speed during the trip in km/h
        maxSpeed:
          type: number
          format: double
          description: Maximum speed reached during the trip in km/h
        spentFuel:
          type: number
          format: double
          description: Amount of fuel consumed during the trip in liters
        startOdometer:
          type: number
          format: double
          description: Odometer reading at trip start
        endOdometer:
          type: number
          format: double
          description: Odometer reading at trip end
        startTime:
          type: string
          format: date-time
          description: Trip start time in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
        endTime:
          type: string
          format: date-time
          description: Trip end time in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
        startPositionId:
          type: integer
          format: int64
          description: ID of the position record at trip start
        endPositionId:
          type: integer
          format: int64
          description: ID of the position record at trip end
        startLat:
          type: number
          format: double
          description: Latitude coordinate at trip start
        startLon:
          type: number
          format: double
          description: Longitude coordinate at trip start
        endLat:
          type: number
          format: double
          description: Latitude coordinate at trip end
        endLon:
          type: number
          format: double
          description: Longitude coordinate at trip end
        startAddress:
          type: string
          maxLength: 512
          description: Human-readable address at trip start location
          nullable: true
        endAddress:
          type: string
          maxLength: 512
          description: Human-readable address at trip end location
          nullable: true
        duration:
          type: integer
          format: int64
          description: Trip duration in milliseconds
        driverId:
          type: integer
          format: int64
          description: ID of the driver who performed this trip
          nullable: true
      required:
        - deviceId
        - distance
        - averageSpeed
        - maxSpeed
        - spentFuel
        - startOdometer
        - endOdometer
        - startTime
        - endTime
        - startPositionId
        - endPositionId
        - startLat
        - startLon
        - endLat
        - endLon
        - duration
...
    LogbookEntryType:
      type: integer
      description: |
        Enum representing the type/classification of a logbook entry.
        Used to categorize trips for business reporting and tax purposes.
        Defaults to BUSINESS (1) if not provided during creation.
      enum: [1, 2]
      default: 1
      x-enum-varnames: [BUSINESS, PRIVATE]
      x-enum-descriptions:
        - "BUSINESS (1): Business-related trip (default)"
        - "PRIVATE (2): Private/personal trip"
      example: 1
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    ApiKey:
      type: http
      scheme: bearer